requirements.txt
aws-cdk-lib==2.147.2
constructs>=10.0.0,<11.0.0
cdk-nag
cdklabs.generative-ai-cdk-constructs